changes for demo

+++ TwitterTopicModeling/Twitter/Models/Tweet.cs
@@ -5,12 +5,12 @@ namespace TwitterTopicModeling.Twitter.Models
     public class Tweet
     {
         [JsonProperty("id")]
-        public int Id {get; set; }
-        
+        public long Id { get; set; }
+
         [JsonProperty("text")]
-        public string Text {get; set; }
+        public string Text { get; set; }
 
         [JsonProperty("user")]
-        public TwitterUser User {get; set; }
+        public TwitterUser User { get; set; }
     }
 }
